+++ PyNucleus/__init__.py
@@ -11,14 +11,12 @@ import pkgutil
 import sys
 
 subpackages = {}
+__all__ = []
 for finder, name, ispkg in pkgutil.iter_modules():
     if ispkg and name.find('PyNucleus_') == 0:
         importName = name[len('PyNucleus_'):]
         module = importlib.import_module(name, 'PyNucleus')
         sys.modules['PyNucleus.'+importName] = module
         subpackages[importName] = module
-        if "__all__" in module.__dict__:
-            names = module.__dict__["__all__"]
-        else:
-            names = [name for name in module.__dict__ if not name.startswith('_')]
+        names = [name for name in module.__dict__ if not name.startswith('_')]
         locals().update({name: getattr(module, name) for name in names})
