+++ README.rst
@@ -32,7 +32,7 @@ Features
 * Nonlocal kernels:
 
   * Finite and infinite horizon
-  * Singularities: fractional, peridynamic, constant kernel
+  * Singularities: fractional, peridynamic, constant, Gaussian kernel
   * spatially variable kernels: variable fractional order and variable coefficients
 
 * Nonlocal assembly (1D and 2D) into dense, sparse and hierarchical matrices
@@ -71,8 +71,39 @@ and open ``docs/build/index.html`` in your browser.
 
 
 
-Prerequisites
-=============
+Possible ways to install and use PyNucleus
+==================================
+
+* Spack install,
+* manual install,
+* Docker container.
+
+Spack install
+-------------
+
+In order to install Spack itself, follow the instructions at https://github.com/spack/spack.
+
+Install PyNucleus and all its dependencies with the command
+
+.. code-block:: shell
+
+   spack install py-pynucleus
+
+To then load PyNucleus
+
+.. code-block:: shell
+
+   spack load py-pynucleus
+
+The examples can be found in the install directory. In order to get there:
+
+.. code-block:: shell
+
+   spack cd -i py-pynucleus
+
+
+Manual installation
+-------------------
 
 In order to install PyNucleus, you will need
 
@@ -83,20 +114,32 @@ In order to install PyNucleus, you will need
 * SuiteSparse,
 * make.
 
-On Debian, Ubuntu etc, the required dependecies can be installed with
+On Debian, Ubuntu etc, the required dependencies can be installed with
 
 .. code-block:: shell
 
    sudo apt-get install python3 mpi-default-bin mpi-default-dev libmetis-dev libparmetis-dev libsuitesparse-dev
 
-Installation
-============
+On MacOS the required dependencies can be installed with
 
-PyNucleus is installed via
+.. code-block:: shell
+
+   brew install python open-mpi
+   brew tap brewsci/num
+   brew install brewsci-metis brewsci-parmetis brewsci-suite-sparse
+
+After cloning the source code, PyNucleus is installed via
 
 .. code-block:: shell
 
    make
+
+The compilation of PyNucleus can be configured by modifying the file `config.yaml` in the root folder.
+This allows for example to set paths for libraries that are installed in non-standard directories.
+The defaults are as follows:
+
+.. literalinclude:: ../config.yaml
+   :language: yaml
 
 If you want to easily modify the source code without re-installing the package every time, and editable install is available as
 
@@ -120,9 +163,9 @@ PyNucleus depends on other Python packages that will be installed automatically:
 
 
 Docker container
-================
+----------------
 
-A Docker container that contains all the required dependecies can be built as well:
+A Docker container that contains all the required dependencies can be built as well:
 
 .. code-block:: shell
 
@@ -144,7 +187,9 @@ or
 Funding
 =======
 
-PyNucleus' development is funded through the MATNIP project (PI: Marta D'Elia) of the LDRD program at Sandia National Laboratories.
+PyNucleus' development is funded through the FOMSI project (PI: Christian Glusa, FY23-FY25) of the LDRD program at Sandia National Laboratories.
+
+PyNucleus' development was previously funded through the MATNIP project (PI: Marta D'Elia, FY20-22).
 
 .. image:: data/matnip.png
    :height: 100px
