+++ multilevelSolver/PyNucleus_multilevelSolver/geometricMG.py
@@ -10,7 +10,7 @@ from __future__ import division
 import logging
 import numpy as np
 from PyNucleus_base.ip_norm import norm_serial as norm
-from PyNucleus_base import getLoggingTimer
+from PyNucleus_base.utilsFem import getLoggingTimer
 from PyNucleus_base import REAL, INDEX, uninitialized
 from PyNucleus_fem import P1_DoFMap
 from PyNucleus_base.linear_operators import LinearOperator
