+++ tests/test_base.py
@@ -10,3 +10,4 @@ from drivers_base import *
 
 from tupleDict import *
 from bitArray import *
+from drift import *
