+++ tests/test_fracLapl.py
@@ -9,15 +9,16 @@
 from __future__ import division
 import numpy as np
 import numpy.linalg
-from PyNucleus.fem import simpleInterval, circle, P1_DoFMap, P2_DoFMap
-from PyNucleus.fem import constant
+from PyNucleus.fem.mesh import simpleInterval, circle
+from PyNucleus.fem.DoFMaps import P1_DoFMap, P2_DoFMap
+from PyNucleus.fem.functions import constant
 from PyNucleus.nl.nonlocalLaplacian import (assembleNonlocalOperator,
                                             
                                             nonlocalBuilder)
 from PyNucleus.nl.clusterMethodCy import H2Matrix
 from PyNucleus.base.myTypes import REAL
 from scipy.special import gamma
-from PyNucleus.nl import getFractionalKernel
+from PyNucleus.nl.kernels import getFractionalKernel
 from PyNucleus.nl.fractionalOrders import (constFractionalOrder,
                                            variableFractionalLaplacianScaling)
 import pytest
