+++ tests/test_nearField.py
@@ -13,10 +13,13 @@ import numpy as np
 from PyNucleus.base.myTypes import REAL, INDEX
 from PyNucleus.base import uninitialized
 from PyNucleus.base.tupleDict import arrayIndexSet
-from PyNucleus.fem import P0_DoFMap, P1_DoFMap, constant
-from PyNucleus.nl import H2Matrix, nonlocalBuilder, getFractionalKernel
+from PyNucleus.fem.DoFMaps import P0_DoFMap, P1_DoFMap
+from PyNucleus.fem.functions import constant
+from PyNucleus.nl import nonlocalBuilder
+from PyNucleus.nl.kernels import getFractionalKernel
 from PyNucleus.nl.nonlocalLaplacian import nearFieldClusterPair
-from PyNucleus.nl.clusterMethodCy import (getDoFBoxesAndCells,
+from PyNucleus.nl.clusterMethodCy import (H2Matrix,
+                                          getDoFBoxesAndCells,
                                           tree_node,
                                           getRefinementParams)
